+++ src/main/java/interface_adapter/schedule/event/AddEventController.java
@@ -15,7 +15,7 @@ public class AddEventController {
         this.addNewEventInteractor = addNewEventInteractor;
     }
 
-    public void postEvent(int projectId, int scheduleId, String eventName, String notes, Date startAt, Date endAt, boolean isAllDay, List<String> userwith) {
+    public void postEvent(long projectId, long scheduleId, String eventName, String notes, Date startAt, Date endAt, boolean isAllDay, List<String> userwith) {
         AddNewEventInputData addNewEventInputData = new AddNewEventInputData(projectId,scheduleId,eventName,notes,startAt,endAt,isAllDay,userwith);
         addNewEventInteractor.postEvent(addNewEventInputData);
     }
