+++ src/main/java/view/JButtonWithFont.java
@@ -4,7 +4,7 @@ import javax.swing.*;
 import java.awt.*;
 
 public class JButtonWithFont extends JButton {
-    private final int defaultSize = 20;
+    private final long defaultSize = 20;
     public JButtonWithFont() {
         super();
         setFont(new Font("Times New Roman", Font.PLAIN, defaultSize));
@@ -16,17 +16,17 @@ public class JButtonWithFont extends JButton {
         setFont(new Font("Times New Roman", Font.PLAIN, defaultSize));
     }
 
-    public JButtonWithFont(String text, int size) {
+    public JButtonWithFont(String text, long size) {
         super(text);
         setFont(new Font("Times New Roman", Font.PLAIN, size));
     }
 
-    public JButtonWithFont(String text, int style, int size) {
+    public JButtonWithFont(String text, long style, long size) {
         super(text);
         setFont(new Font("Times New Roman", style, size));
     }
 
-    public JButtonWithFont(String text, String name, int style, int size) {
+    public JButtonWithFont(String text, String name, long style, long size) {
         super(text);
         setFont(new Font(name, style, size));
     }
